backend/requirements-dev.txt
# Dependências para desenvolvimento e testes
asgiref==3.9.1
Django==5.2.6
sqlparse==0.5.3
tzdata==2025.2

# Dependências para testes
factory-boy==3.3.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
coverage==7.3.2

# Dependências para API (quando implementar DRF)
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Dependências para desenvolvimento
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Dependências para qualidade de código
black==23.9.1
flake8==6.1.0
isort==5.12.0
mypy==1.6.1
django-stubs==4.2.7
